refactored client fetch logic
