+++ client/posts.tsx
@@ -1,5 +1,5 @@
 import { PostEntity, PostRequestData } from "@/server/posts";
-import { fetchData, FetchEndpoints, FetchMethods } from "./core";
+import { request, FetchEndpoints, FetchMethods } from "@/lib/api";
 
 interface UserRequestData {
     userId: string;
@@ -10,12 +10,12 @@ export interface PostResponse {
     message?: string,
 }
 
-export const createPost = (content: PostRequestData, callback?: () => void) =>
-    fetchData<PostResponse, PostRequestData>(
+export const createPost = async (content: PostRequestData, callback?: () => void) =>
+    await request<PostResponse, PostRequestData>(
         FetchEndpoints.POSTS,
         { method: FetchMethods.POST, body: content, callback });
 
-export const deletePost = (postId: string, userId: string, callback?: () => void) =>
-    fetchData<PostResponse, UserRequestData>(
+export const deletePost = async (postId: string, userId: string, callback?: () => void) =>
+    await request<PostResponse, UserRequestData>(
         FetchEndpoints.POSTS,
         { method: FetchMethods.DELETE, body: { userId }, query: [postId], callback });
